Final commit for MemeMe v1
Unable to get the captions to automatically be placed within the
image onscreen and also unable to capture only the image and
captions on screen, eliminating the white lettterboxing effect on
the final image.
This is all DESPITE spending much more than 10x the time on
this one feature as was spent on the entire rest of the project.